requirements.txt
absl-py==0.9.0
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
data==0.4
decorator==4.4.1
funcsigs==1.0.2
future==0.18.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
grpcio==1.26.0
idna==2.8
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
kiwisolver==1.1.0
line-profiler==3.0.2
Markdown==3.2
matplotlib==3.1.3
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.4.3
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
shutilwhich==1.1.0
six==1.14.0
snakeviz==2.0.1
tempdir==0.7.1
tensorboard==2.1.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
seaborn
